+++ app/src/main/java/com/todobom/opennotescanner/helpers/AboutFragment.java
@@ -1,8 +1,10 @@
 package com.todobom.opennotescanner.helpers;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.DialogFragment;
+import android.view.Display;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -17,6 +19,7 @@ import us.feras.mdv.MarkdownView;
  */
 public class AboutFragment extends DialogFragment {
 
+    private static final String APP_LINK = "https://goo.gl/2JwEPq";
     private Runnable mRunOnDetach;
 
     public AboutFragment() {
@@ -46,28 +49,28 @@ public class AboutFragment extends DialogFragment {
 
         markdownView.loadMarkdownFile("file:///android_asset/" + getString(R.string.about_filename));
 
-        /*
+        Display display = getActivity().getWindowManager().getDefaultDisplay();
+        android.graphics.Point size = new android.graphics.Point();
+        display.getRealSize(size);
 
-        StringBuilder aboutBuffer=new StringBuilder();
-        BufferedReader in;
-        String aboutString="";
-        try {
-            InputStream aboutStream=getContext().getAssets().open(getString(R.string.about_filename));
-            in = new BufferedReader(new InputStreamReader(aboutStream, "UTF-8"));
-            while ((aboutString=in.readLine()) != null) {
-                aboutBuffer.append(aboutString + "\n");
+        View aboutContainer = view.findViewById(R.id.about_container);
+        ViewGroup.LayoutParams params = aboutContainer.getLayoutParams();
+        params.height = (int) (size.y * .9);
+        params.width = (int) (size.x * .9);
+        aboutContainer.setLayoutParams(params);
+
+        View about_shareapp = view.findViewById(R.id.about_shareapp);
+        about_shareapp.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                String shareBody = getString(R.string.share_app_body) + APP_LINK;
+                Intent sharingIntent = new Intent(android.content.Intent.ACTION_SEND);
+                sharingIntent.setType("text/plain");
+                sharingIntent.putExtra(android.content.Intent.EXTRA_SUBJECT, getString(R.string.share_app_subject));
+                sharingIntent.putExtra(android.content.Intent.EXTRA_TEXT, shareBody);
+                startActivity(Intent.createChooser(sharingIntent, getString(R.string.share_app_using)));
             }
-            aboutString = aboutBuffer.toString();
-            in.close();
-
-            markdownView.setStringContent(aboutString);
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        /* */
-
+        });
     }
 
     @Override