+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+git HEAD
+--------
+
+* better images in color mode
+* enhancements on main screen interface
+* flash state persist between activities changes
+
 version 1.0.19
 --------------